Reorganize react components folder structure
