+++ Sources/frontend-react-js/src/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { StreamdeckSocketSettings, defaultStreamdeckSocketSettings, useStreamdeckWebsocket } from "./api/StreamdeckWebsocket";
+import { StreamdeckSocketSettings, defaultStreamdeckSocketSettings, useStreamdeckWebsocket } from "./api/Streamdeck/StreamdeckWebsocket";
 import ButtonSettings from "./areas/ButtonSettings";
 import IdLookup from "./areas/IdLookup";
 import RightSidebar from "./areas/RightSidebar";
