+++ Sources/frontend-react-js/src/api/DcsBios/ControlReferenceInterface.tsx
