+++ Sources/frontend-react-js/src/api/DcsBios/FlattenModuleControlsJson.tsx
