+++ Sources/frontend-react-js/src/components/IdLookupManager.module.css
