+++ Sources/frontend-react-js/src/components/SearchBar.module.css
