+++ Sources/frontend-react-js/src/forms/SendCommand.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, useEffect, useState } from "react";
-import StreamdeckApi from "../api/StreamdeckApi";
+import StreamdeckApi from "../api/Streamdeck/StreamdeckApi";
 
 export interface SendCommandSettings {
     send_address: string;
