+++ Sources/frontend-react-js/src/modals/Backdrop.module.css
@@ -1,0 +1,17 @@
+.backdrop {
+  position: absolute;
+  z-index: 1;
+  background-color: rgba(0, 0, 0, 0.4);
+  width: 100%;
+  height: 100vh;
+  top: 0;
+  left: 0;
+}
+
+.local{
+  position: absolute;
+}
+
+.global{
+  position: fixed;
+}
