+++ Sources/frontend-react-js/src/modals/SelectedControlRef.module.css
@@ -11,21 +11,3 @@
   top: 20vh;
   left: calc(50% - 15rem);
 }
-
-.backdrop {
-  position: absolute;
-  z-index: 1;
-  background-color: rgba(0, 0, 0, 0.4);
-  width: 100%;
-  height: 100vh;
-  top: 0;
-  left: 0;
-}
-
-.local{
-  position: absolute;
-}
-
-.global{
-  position: fixed;
-}
